satellite/admin: add coupon deletion (#3893)
