+++ satellite/admin/server.go
@@ -71,6 +71,7 @@ func NewServer(log *zap.Logger, listener net.Listener, db DB, config Config) *Se
 	server.mux.HandleFunc("/api/user/{useremail}", server.userInfo).Methods("GET")
 	server.mux.HandleFunc("/api/coupon", server.addCoupon).Methods("POST")
 	server.mux.HandleFunc("/api/coupon/{couponid}", server.couponInfo).Methods("GET")
+	server.mux.HandleFunc("/api/coupon/{couponid}", server.deleteCoupon).Methods("DELETE")
 	server.mux.HandleFunc("/api/project/{project}/limit", server.getProjectLimit).Methods("GET")
 	server.mux.HandleFunc("/api/project/{project}/limit", server.putProjectLimit).Methods("PUT", "POST")
 	server.mux.HandleFunc("/api/project/{project}", server.deleteProject).Methods("DELETE")
